Update rvy_init.php
PostMeta Failsafe: to avoid the possibility of accidental clearance, Featured Image removal is not revisioned, until further testing. API filter "revisionary_removable_meta_fields" available for experimental usage with specified meta keys.

+++ rvy_init.php
@@ -614,7 +614,9 @@ function revisionary_copy_meta_field( $meta_key, $from_post_id, $to_post_id, $mi
 			)
 		) {
 			update_post_meta($to_post_id, $meta_key, $source_meta->meta_value);
-		} elseif ($mirror_empty) {
+
+		} elseif ($mirror_empty && in_array($meta_key, apply_filters('revisionary_removable_meta_fields', [], $to_post_id))) {
+			// Disable postmeta deletion until further testing
 			delete_post_meta($to_post_id, $meta_key);
 		}
 	}
